% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DENDRO_PLOT_FUNCTIONS.R
\name{find.outliers}
\alias{find.outliers}
\title{Automatically identify outliers in dendrometer band time series}
\usage{
find.outliers(ts.data, sd.lim = 3)
}
\arguments{
\item{ts.data}{A dataframe of a time series of a single tree in a year.
Must have column variables \emph{DBH_TRUE} (numeric) and \emph{DOY} (integer),
as well as other designations.}

\item{params}{dataframe vector}

\item{day}{integer vector}

\item{par.names}{A character vector of the column names of the \emph{params} vector argument.
This is used to pull out actual parameters from other information.}
}
\description{
Identifies outliers that are beyond the determined standard
deviation of the residuals. Assumes residuals are from a fit model.
}
