% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DENDRO_BASE_FUNCTIONS.R
\name{gap2dbh}
\alias{gap2dbh}
\title{Gets dbh from gap width given org.dbh.}
\usage{
gap2dbh(gap.width, org.dbh, unit = "cm")
}
\arguments{
\item{gap.width}{Numeric vector of gap width measurements}

\item{org.dbh}{Numeric scalar of the original dbh of the tree (when bands were installed)}

\item{unit}{Character string denoting the unit that the dbh is in. Defaults to "cm".}
}
\value{
Numeric vector of DBH_TRUE values
}
\description{
Gets dbh from gap width given org.dbh.
}
