% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DENDRO_BASE_FUNCTIONS.R
\name{get.alt.a}
\alias{get.alt.a}
\title{Gets the starting diameter from the year before}
\usage{
get.alt.a(param.tab)
}
\arguments{
\item{param.tab}{data.frame of parameter values for every year for a particular TREE_ID}
}
\value{
Numeric vector of diameter values (or codes).
}
\description{
Gets the starting diameter from the year before
}
