% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DENDRO_BASE_FUNCTIONS.R
\name{get.extra.metrics}
\alias{get.extra.metrics}
\title{Extract extra growth metrics.}
\usage{
get.extra.metrics(param.table, Dendro.split, par.names = c("L", "K",
  "doyip", "r", "theta", "a", "b", "r.squared", "ts.sd", "alt.a"),
  OUTPUT.folder = "OUTPUT",
  param.table.name = "Param_table_complete.csv",
  Dendro.split.name = "Dendro_split.Rdata",
  Dendro.data.name = "Dendro_data.Rdata",
  Quantile.hull.name = "Quantile.hull.Rdata")
}
\arguments{
\item{param.table}{The output data.frame from the optimization @seealso get.optimized.dendro}

\item{Dendro.split}{A list vector where every entry is a separate time series of a band on a stem in a year.}

\item{par.names}{A character vector of the column names of the \emph{params} vector argument.
This is used to pull out actual parameters from other information.}

\item{OUTPUT.folder}{Character string for the target output folder.}

\item{param.table.name}{Character string for the parameter table output.}

\item{Dendro.data.name}{Character string for the data object, Dendro.complete, that is a complete data.frame table output.}

\item{Quantile.hull.name}{Character string for the data object, QH.Rdata, that contains the Quantile Hull results.}
}
\value{
Returns nothing.

nothing
}
\description{
Fits a Quantile Hull to the data and extracts extra growth and phenology metrics from dendrometer band time series.
}
\examples{
none
}
