% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DENDRO_BASE_FUNCTIONS.R
\name{get.optimized.dendro}
\alias{get.optimized.dendro}
\title{Fit the LG5 function}
\usage{
get.optimized.dendro(INPUT.dendro, no.neg.growth = TRUE, cutoff = 9,
  par.names = c("L", "K", "doyip", "r", "theta", "a", "b", "r.squared",
  "ts.sd"), OUTPUT.folder = "OUTPUT",
  param.table.name = "Param_table.csv",
  Dendro.data.name = "Dendro_data.Rdata",
  Dendro.split.name = "Dendro_data_split.Rdata")
}
\arguments{
\item{INPUT.dendro}{A data.frame of a specific form (see @details)}

\item{no.neg.growth}{logical denoting whether to skip any time series with negative trends (from a call to \emph{lm()})}

\item{cutoff}{Integer denoting the minimum sample size required to perform the optimization call.}

\item{par.names}{A character vector of the column names of the \emph{params} vector argument.
This is used to pull out actual parameters from other information.}

\item{OUTPUT.folder}{Character string for the target output folder.}

\item{param.table.name}{Character string for the parameter table output.}

\item{Dendro.data.name}{Character string for the data object, Dendro.complete, that is a complete data.frame table output.}

\item{Dendro.split.name}{Character string for the data object, Dendro.split, that is a list vector where
every entry is a separate time series of a band on a stem in a year.}
}
\value{
Returns nothing.
}
\description{
Estimate LG5 fit, structure output, and get summary statistics for dendrometer band time series.
}
\examples{
none
}
