% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DENDRO_BASE_FUNCTIONS.R
\name{get.QH.resid}
\alias{get.QH.resid}
\title{Get quantile hull 'residuals'}
\usage{
get.QH.resid(params, doy, dbh, log.rate = FALSE)
}
\arguments{
\item{params}{Numeric vector of parameters.}

\item{doy}{Day of the Year}

\item{dbh}{Sequence of DBH values}

\item{log.rate}{Logical whether rates are logged}
}
\value{
Numeric vector of residuals
}
\description{
Get quantile hull 'residuals'
}
