% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DENDRO_BASE_FUNCTIONS.R
\name{get_dcrit}
\alias{get_dcrit}
\title{Gets the day of the year for any dbh and paramater combination}
\usage{
get_dcrit(dbh, params)
}
\arguments{
\item{dbh}{Numeric scalar for diameter at breast height (translated from the GAP_WIDTH)}

\item{params}{Numeric vector of the parameter values of the LG5}
}
\value{
Returns day of the year for the input dbh and params.
}
\description{
Gets the day of the year for any dbh and paramater combination
}
