% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DENDRO_BASE_FUNCTIONS.R
\name{gettruedbh}
\alias{gettruedbh}
\title{Corrects for the chord}
\usage{
gettruedbh(gw1, gw2, dbh1)
}
\arguments{
\item{gw1}{Numeric gap width at time 1}

\item{gw2}{Numeric gap width at time 2}

\item{dbh1}{DBH at time 1}
}
\value{
Numeric scalar for the corrected DBH
}
\description{
Corrects for the chord
}
