% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{INPUT.data}
\alias{INPUT.data}
\title{Sample Dendrometer band dataset}
\format{A data frame with 781 rows and 15 variables:
\describe{
  \item{SITE}{Location of observations}
  \item{PROJECT}{character string, name of the project for the big database.}
  ...
}}
\usage{
INPUT.data
}
\description{
Dendrometer band time series for three trees over multiple years at SERC
}
\keyword{datasets}
