% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{INPUT.dendro}
\alias{INPUT.dendro}
\title{Sample Dendrometer band dataset}
\format{A data frame with  1589 rows and 16 variables:
\describe{
  \item{SITE}{Location of observations}
  \item{PROJECT}{charcter string, name of the project for the big database.}
  ...
}}
\usage{
INPUT.dendro
}
\description{
Dendrometer band time series for trees across years at SERC (a  sample)
}
\keyword{datasets}
