% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DENDRO_BASE_FUNCTIONS.R
\name{lg5.deriv}
\alias{lg5.deriv}
\title{Derivative of LG5 function}
\usage{
lg5.deriv(paras, doy, growth = 1, shift = 0.5)
}
\arguments{
\item{paras}{Numeric vector of parameters for the lg5 function}

\item{doy}{Integer vector of days of the year dbh values collected.}

\item{growth}{Unclear.}

\item{shift}{Numeric scalar that gives half the size of the window the derivitive is taken over (in days).}
}
\value{
returns numeric vectore of derivatives
}
\description{
This function takes the numerical derivative.
Default values return a single day of growth. Using the 'growth'
argument, the derivative can be returned, scaled by annual growth.
}
