% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DENDRO_BASE_FUNCTIONS.R
\name{lg5.pred.a}
\alias{lg5.pred.a}
\title{Returns starting and stopping diameter values for a year for a tree}
\usage{
lg5.pred.a(a, params, doy, asymptote = "both")
}
\arguments{
\item{a}{Numeric scalar or vector of 2 elements with starting and stopping values}

\item{params}{Numeric vector of parameters}

\item{doy}{Integer (or numeric) vector of days of the year for which the curve was fit.}

\item{asymptote}{which side to get value for (should be removed).}
}
\value{
Numeric scalar (or vector of 2 scalars) for starting and stoping sizes
}
\description{
Returns starting and stopping diameter values for a year for a tree
}
