% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DENDRO_PLOT_FUNCTIONS.R
\name{make.dendro.plot.tree}
\alias{make.dendro.plot.tree}
\title{Plotting dendrometer Band Time Series for a single tree over years}
\usage{
make.dendro.plot.tree(Dendro.ind, param.tab)
}
\arguments{
\item{Dendro.ind}{A dataframe of a time series of a single tree over multiple years and bands.
Must have column variables \emph{DBH_TRUE} (numeric), \emph{DOY} (integer), and \emph{YEAR} (integer).}

\item{param.tab}{dataframe of fit parameters}
}
\description{
Plots an extended time series of dendrometer band measurements (translated into DBH) with
fits, outliers, band position movements, slippage corrections, etc. Can be used for presentation
of single trees, or for diagnostics and fit assessments.
}
