% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DENDRO_PLOT_FUNCTIONS.R
\name{make.dendro.plot.ts}
\alias{make.dendro.plot.ts}
\title{Plotting Dendrometer Band Time Series}
\usage{
make.dendro.plot.ts(ts.data, params = NULL, day = seq(365),
  outlier = TRUE, par.names = c("L", "K", "doyip", "r", "theta", "a",
  "b", "alt.a"))
}
\arguments{
\item{ts.data}{A dataframe of a time series of a single tree in a year.
Must have column variables \emph{DBH_TRUE} (numeric) and \emph{DOY} (integer).}

\item{params}{dataframe vector (i.e., there are names associated with the vector elements)}

\item{day}{integer vector}

\item{outlier}{logical}

\item{par.names}{A character vector of the column names of the params vector. This is used to pull out actual parameters from other information.}
}
\value{
A plot of the dbh and doy of a single band in a year, with, optionally, a fitted line from the optimize output and outlier denotion in red.
}
\description{
Plotting Dendrometer Band Time Series
}
\seealso{
\code{\link{get.params}}, which creates Dendro.split, a list vector containing time-series dataframes for every tree, year, band.
}
