% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DENDRO_BASE_FUNCTIONS.R
\name{max.growth.day}
\alias{max.growth.day}
\title{Find day of the year of maximum growth}
\usage{
\method{max}{growth.day}(params)
}
\arguments{
\item{params}{Numeric vector of parameters.}
}
\value{
Numeric scalar of the fastest day.
}
\description{
Find day of the year of maximum growth
}
