% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DENDRO_BASE_FUNCTIONS.R
\name{max.growth.rate}
\alias{max.growth.rate}
\title{Find rate of maximum growth}
\usage{
\method{max}{growth.rate}(params)
}
\arguments{
\item{params}{Numeric vector of parameters.}
}
\value{
Numeric scalar of the fastest growth rate
}
\description{
Find rate of maximum growth
}
