% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DENDRO_BASE_FUNCTIONS.R
\name{outer.hull}
\alias{outer.hull}
\title{Fit Quantile Hull to subset of time series data}
\usage{
outer.hull(dbh, doy, params, quant = 0.8, resid.sd = 0.02)
}
\arguments{
\item{dbh}{Numeric vector of "TRUE_DBH" values}

\item{doy}{Integer vector of days of the year dbh values collected.}

\item{params}{Data.frame (vector with names) containing parameters.}

\item{quant}{Numeric scaler denoting the quantile threshold above which the hull will be estimated.}

\item{resid.sd}{Numeric scaler that constrains the optimization algorithm in \emph{optim()}.}
}
\value{
List with output from the optim function and other metrics.
}
\description{
Fit Quantile Hull to subset of time series data
}
\examples{
none
}
