% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DENDRO_PLOT_FUNCTIONS.R
\name{plot.dendro.ts}
\alias{plot.dendro.ts}
\title{Plotting Dendrometer Band Time Series}
\usage{
\method{plot}{dendro.ts}(ts.data, params = NULL, day = seq(365),
  outlier = TRUE, par.names = c("L", "K", "doyip", "r", "theta", "a",
  "b", "alt.a"))
}
\arguments{
\item{ts.data}{A dataframe of a time series of a single tree in a year.
Must have column variables \emph{DBH_TURE} (numeric) and \emph{DOY} (integer).}

\item{params}{dataframe vector}

\item{day}{integer vector}

\item{outlier}{logical}

\item{par.names}{A character vector of the column names of the params vector. This is used to pull out actual parameters from other information.}
}
\value{
A plot of the dbh and doy of a single band in a year, with, optionally, a fitted line from the optimize output and outlier denotion in red.
}
\description{
Plotting Dendrometer Band Time Series
}
\examples{
plot.dendro.ts(data.frame(DBH_TRUE = c(10.1, 10.1., 10.2, 10.3), DOY = c(70, 100, 140, 160))
}
\seealso{
\code{\link{get.param}} which creates Dendro.split, a list vector containing time-series dataframes for every tree, year, band.
}
