% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DENDRO_BASE_FUNCTIONS.R
\name{pred.doy}
\alias{pred.doy}
\title{Predicts the day of the year given a diameter and parameter values}
\usage{
pred.doy(params, a, diam.given = 0)
}
\arguments{
\item{params}{Numeric vector of parameter values.}

\item{a}{Numeric scalar of diameter for which a doy is wanted}

\item{diam.given}{Not sure}
}
\value{
returns day of year
}
\description{
Predicts the day of the year given a diameter and parameter values
}
