% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DENDRO_BASE_FUNCTIONS.R
\name{get.optimized.dendro}
\alias{get.optimized.dendro}
\title{Fit the LG5 function}
\usage{
get.optimized.dendro(INPUT.data, no.neg.growth = TRUE, cutoff = 9,
  units = "cm", par.names = c("L", "K", "doyip", "r", "theta", "a",
  "b", "r.squared", "ts.sd"), OUTPUT.folder = ".",
  param.table.name = "Param_table.csv",
  Dendro.data.name = "Dendro_data.Rdata",
  Dendro.split.name = "Dendro_data_split.Rdata")
}
\arguments{
\item{INPUT.data}{A data.frame of a specific form (see @details)}

\item{no.neg.growth}{logical denoting whether to skip any time series with negative trends (from a call to \emph{lm()})}

\item{cutoff}{Integer denoting the minimum sample size required to perform the optimization call.}

\item{units}{Character string denoting the units that the dbh is in. Defaults to "cm".}

\item{par.names}{A character vector of the column names of the \emph{params} vector argument.
This is used to pull out actual parameters from other information.}

\item{OUTPUT.folder}{Character string for the target output folder.}

\item{param.table.name}{Character string for the parameter table output.}

\item{Dendro.data.name}{Character string for the data object, Dendro.complete, that is a complete data.frame table output.}

\item{Dendro.split.name}{Character string for the data object, Dendro.split, that is a list vector where
every entry is a separate time series of a band on a stem in a year.}
}
\value{
Objects Saves four files to the OUTPUT folder as named above and collected stems.
}
\description{
This is the wrapper for \emph{get.params()} which uses base \emph{optim()} (with two sequential methods) to
estimate the parameters of the logistic function. This follows McMahon and Parker 2014. This function organizes output, skips removed datasets, fits a linear model to detect negative or non-significant growth, eliminates time series with small sample sizes, and removes identified outliers (see @identify.outliers ).
}
