% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DENDRO_PLOT_FUNCTIONS.R
\name{id.outliers}
\alias{id.outliers}
\title{Identify outliers in dendrometer band time series}
\usage{
id.outliers(ts.data, params, day = seq(365), par.names = c("L", "K",
  "doyip", "r", "theta", "a", "b", "alt.a"))
}
\arguments{
\item{ts.data}{A dataframe of a time series of a single tree in a year.
Must have column variables \emph{DBH_TRUE} (numeric) and \emph{DOY} (integer), as well as other designations.}

\item{params}{dataframe vector}

\item{day}{integer vector}

\item{par.names}{A character vector of the column names of the \emph{params} vector argument.
This is used to pull out actual parameters from other information.}
}
\value{
none
}
\description{
An interactive figure for the identification and designation of slipped bands, outliers, and problematic bands.
}
\examples{
none
}
