% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DENDRO_BASE_FUNCTIONS.R
\name{lg5.pred}
\alias{lg5.pred}
\title{Predicts the dbh from doy and a parameter set}
\usage{
lg5.pred(params, doy)
}
\arguments{
\item{params}{Numeric vector of parameter values}

\item{doy}{Integer vector of days of the year}
}
\value{
Numeric scalar or vector of dbh values
}
\description{
Predicts the dbh from doy and a parameter set
}
